🌱 feat: 장소, 전체 일정 통계 조회 api 구현

+++ apps/backend/src/main/java/com/calendarbox/backend/analytics/dto/response/DayHourScheduleDistribution.java
@@ -3,6 +3,6 @@ package com.calendarbox.backend.analytics.dto.response;
 public record DayHourScheduleDistribution(
         Integer dayOfWeek,
         Integer hourOfDay,
-        Integer scheduleCount
+        Long scheduleCount
 ) {
 }
