+++ apps/backend/src/main/java/com/calendarbox/backend/analytics/dto/response/PeopleStatItem.java
@@ -5,8 +5,8 @@ public record PeopleStatItem(
         String name,
         Integer meetCount,
         Long totalDurationMin,
-        Long avgDurationMin,
+        Double avgDurationMin,
         Long totalAmount,
-        Long avgAmount
+        Double avgAmount
 ) {
 }
