+++ apps/backend/src/main/java/com/calendarbox/backend/analytics/dto/response/PersonMonthlyExpenseSummary.java
@@ -6,7 +6,7 @@ public record PersonMonthlyExpenseSummary(
         LocalDateTime month,
         Long personId,
         String personName,
-        Double totalAmount,
+        Long totalAmount,
         Double avgAmount,
         Integer sharedScheduleCount
 ) {
